Remove old prototyping directories and rename the latest.
