+++ prototyping/aliases.py
@@ -1,21 +1,35 @@
-"""A collection of aliases for `str` and `Any` that make type annotations
+"""A collection of aliases for `str`, `Any`, etc. that make type annotations
 more readable.
 
 While we could consider using these in the real codebase, the purpose of this
 file isn't to propose that, and I am definitely not in favor of going all the
-way to `typing.NewType`, as these pretty much all need to be appear in
-public interfaces where the newtype would be a big hassle.
+way to `typing.NewType`, as these pretty much all need to be appear in public
+interfaces where the newtype would be a big hassle.
 """
 
 from __future__ import annotations
 
-from collections.abc import Mapping
+__all__ = (
+    "CollectionDocumentation",
+    "CollectionName",
+    "CollectionPattern",
+    "ColumnName",
+    "DatasetTypeName",
+    "DatasetTypePattern",
+    "DimensionElementName",
+    "DimensionName",
+    "GetParameter",
+    "InMemoryDataset",
+    "DatastoreTableName",
+    "StorageClassName",
+    "StorageURI",
+    "TransferMode",
+)
+
 from typing import Any, TypeAlias
 
-from lsst.resources import ResourcePath
-
-CollectionName: TypeAlias = str
 CollectionDocumentation: TypeAlias = str
+CollectionName: TypeAlias = str
 CollectionPattern: TypeAlias = Any
 ColumnName: TypeAlias = str
 DatasetTypeName: TypeAlias = str
@@ -24,35 +38,7 @@ DimensionElementName: TypeAlias = str
 DimensionName: TypeAlias = str
 GetParameter: TypeAlias = str
 InMemoryDataset: TypeAlias = Any
-OpaqueTableName: TypeAlias = str
+DatastoreTableName: TypeAlias = str
 StorageClassName: TypeAlias = str
-FormatterName: TypeAlias = str
-
-
-JournalPathMap = Mapping[str, Mapping[str, ResourcePath]]
-"""A nested mapping indexed first by Datastore name, then RUN collection name,
-with signed-if-needed URLs for journal files that Datastore should create when
-starting a transaction and delete when a transaction finishes successfully.
-"""
-
-OpaqueTableValues = Any
-"""Per-table, per-dataset opaque values.
-
-This represents either:
-
-- a single row in an opaque table for which the dataset UUID is the only
-  primary key, which could just be a `dict` or even a (named) `tuple`.
-
-- multiple rows in an opaque table for which the dataset UUID is part of a
-  compound primary key (e.g. when component is another key), which could be a
-  `list` or `dict` of `dict` or (named) `tuple` rows.
-
-In both cases, the UUID does not need to be included in the rows, as it's
-always held by some outer data structure.
-
-These need to be built-ins or pydantic models so we can serialize them
-directly.
-
-For datastores that use signed URIs, each row should always hold the
-original unsigned URI and may optionally hold a signed URL as well.
-"""
+StorageURI: TypeAlias = tuple[str, str]  # Datastore name, possibly-relative URI
+TransferMode: TypeAlias = str
