+++ prototyping/removal_transaction.py
